Move to sample.py
